refactor(baselines): consolidate path canonicalization into path-canon (refs #3345)
Fold the permissive canonicalizeBaselinePath coercer from the separate
canonicalize-path.js module into path-canon.js so a single module exposes
both the strict canonicalise/assertCanonical helpers and the permissive
coerce helper. refresh-service.js and the writer-assertion path now import
from one place, removing the cross-referencing indirection between the two
baseline path helpers.
